move private structure to private header, Use explicit values for the constructor
