+++ libs/indibase/property/indiproperty_p.h
@@ -1,0 +1,40 @@
+/*******************************************************************************
+  Copyright(c) 2011 Jasem Mutlaq. All rights reserved.
+
+ This library is free software; you can redistribute it and/or
+ modify it under the terms of the GNU Library General Public
+ License version 2 as published by the Free Software Foundation.
+
+ This library is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ Library General Public License for more details.
+
+ You should have received a copy of the GNU Library General Public License
+ along with this library; see the file COPYING.LIB.  If not, write to
+ the Free Software Foundation, Inc., 51 Franklin Street, Fifth Floor,
+ Boston, MA 02110-1301, USA.
+*******************************************************************************/
+
+#pragma once
+
+#include "indibase.h"
+
+namespace INDI
+{
+
+class BaseDevice;
+class PropertyPrivate
+{
+public:
+    void *property = nullptr;
+    BaseDevice *baseDevice = nullptr;
+    INDI_PROPERTY_TYPE type = INDI_UNKNOWN;
+    bool registered = false;
+    bool dynamic = false;
+
+    PropertyPrivate(void *property, INDI_PROPERTY_TYPE type);
+    virtual ~PropertyPrivate();
+};
+
+}
